Rectangle,Area to long

+++ src/homework48/Rectangle.java
@@ -3,12 +3,12 @@ package homework48;
 public class Rectangle {
     public final int length;
     public final int width;
-    private final int area;
+    private final long area;
 
     public Rectangle(int a, int b) {
         length = Math.max(a, b);
         width = Math.min(a, b);
-        area = a * b;
+        area = (long) a * (long) b;
 
     }
 
@@ -20,7 +20,7 @@ public class Rectangle {
         return width;
     }
 
-    public int getArea() {
+    public long getArea() {
         return area;
     }
